[PLAT-4424][xCluster] Backend support to restart the whole replication config
Summary:
This diff adds backend support to restart an xCluster config. Restarting an xCluster config is combining a DeleteXClusterConfig task followed by a CreateXClusterConfig task, however, it keeps the xCluster config entry in the DB and resets its state.
When it tries to set up replication, it will automatically determine if bootstrap is required and does so. Because of that, to restart an xCluster config containing ysql tables, all the tables in a database must be included in the config, otherwise, the restart operation fails with a message of "For YSQL tables, all the tables in a keyspace must be selected".

To trigger a restart on an xCluster config, the user can make the following request (Please make sure to fill the variables with the correct values):
```
curl --location --request POST '<Platform-address>/api/v1/customers/<customer-uuid>/xcluster_configs/<xCluster-config-uuid>' \
--header 'Content-Type: application/json' \
--header 'X-AUTH-YW-API-TOKEN: <auth-token>' \
--data-raw '{
   "tables":[],
   "bootstrapParams": {
      "backupRequestParams": {
        "storageConfigUUID": <storage-config-uuid>
      }
    }
}'
```
Currently, tables must be empty. It is added for later improvement of restarting the replication for one table. The specified storage config will be used in case bootstrap is required.

E.g.:
```
curl --location --request POST 'http://192.168.56.102:9000/api/v1/customers/f33e3c9b-75ab-4c30-80ad-cba85646ea39/xcluster_configs/e55b1faa-77f1-428a-bbe4-d82837429417' \
--header 'Content-Type: application/json' \
--header 'X-AUTH-YW-API-TOKEN: 2894987a-4fdf-47cf-95a2-be3f2039cf79' \
--data-raw '{
   "tables":[],
   "bootstrapParams": {
      "backupRequestParams": {
        "storageConfigUUID": "cb5e0444-3f36-486d-aa07-20e10c37b382"
      }
    }
}'
```

This diff also makes changes to the DeleteXClusterConfig subtasks so it synchronizes the Platform state with DB before performing the deletion of the replication group and thus makes the clean-up more reliable.

Test Plan:
  - Created xCluster configs with and without bootstrapping
  - Did different xCluster operations such as pause/enable, add/remove table, and rename
  - Restarted the configs and made sure it works

Reviewers: sanketh, nsingh, sb-yb, nbhatia, amalyshev, spotachev

Reviewed By: spotachev

Subscribers: jenkins-bot, sanketh, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D19034

+++ managed/src/main/java/com/yugabyte/yw/commissioner/tasks/subtasks/xcluster/DeleteReplication.java
@@ -15,6 +15,7 @@ import org.yb.WireProtocol;
 import org.yb.client.DeleteUniverseReplicationResponse;
 import org.yb.client.MasterErrorException;
 import org.yb.client.YBClient;
+import org.yb.master.CatalogEntityInfo;
 
 @Slf4j
 public class DeleteReplication extends XClusterConfigTaskBase {
@@ -50,6 +51,8 @@ public class DeleteReplication extends XClusterConfigTaskBase {
     XClusterConfig xClusterConfig = getXClusterConfigFromTaskParams();
 
     if (xClusterConfig.targetUniverseUUID == null) {
+      xClusterConfig.setReplicationSetupDone(
+          xClusterConfig.getTables(), false /* replicationSetupDone */);
       log.info("Skipped {}: the target universe is destroyed", getName());
       return;
     }
@@ -62,9 +65,15 @@ public class DeleteReplication extends XClusterConfigTaskBase {
     String targetUniverseCertificate = targetUniverse.getCertificateNodetoNode();
     try (YBClient client =
         ybService.getClient(targetUniverseMasterAddresses, targetUniverseCertificate)) {
-      // Catch the `Universe replication NOT_FOUND` exception, and because it already does not
-      // exist, the exception will be ignored.
-      try {
+      // Sync the state for the tables in the xCluster config.
+      CatalogEntityInfo.SysClusterConfigEntryPB clusterConfig =
+          getClusterConfig(client, targetUniverse.universeUUID);
+      boolean replicationGroupExists =
+          syncReplicationSetUpStateForTables(
+              clusterConfig, xClusterConfig, xClusterConfig.getTables());
+
+      // If replication group exists, delete it from the cluster config of the target universe.
+      if (replicationGroupExists) {
         DeleteUniverseReplicationResponse resp =
             client.deleteUniverseReplication(
                 xClusterConfig.getReplicationGroupName(), ignoreErrors);
@@ -81,22 +90,26 @@ public class DeleteReplication extends XClusterConfigTaskBase {
                   "Failed to delete replication for XClusterConfig(%s): %s",
                   xClusterConfig.uuid, resp.errorMessage()));
         }
-      } catch (MasterErrorException e) {
-        // If it is not `Universe replication NOT_FOUND` exception, rethrow the exception.
-        if (!e.getMessage().contains("NOT_FOUND[code 1]: Universe replication")) {
-          throw new RuntimeException(e);
-        }
+
+        // After the RPC call, the corresponding stream ids on the source universe will be deleted
+        // as well.
+        xClusterConfig
+            .getTablesById(xClusterConfig.getTableIdsWithReplicationSetup())
+            .forEach(
+                tableConfig -> {
+                  tableConfig.streamId = null;
+                });
+        xClusterConfig.update();
+      } else {
         log.warn(
-            "XCluster config {} does not exist on the target universe, NOT_FOUND exception "
-                + "occurred in deleteUniverseReplication RPC call is ignored: {}",
-            xClusterConfig.uuid,
-            e.getMessage());
+            "XCluster config {} does not exist on the target universe, RPC to delete the "
+                + "replication group will not be called",
+            xClusterConfig.uuid);
       }
 
-      // Set status of the xCluster config to `Deleted` if required.
-      if (!isInMustDeleteStatus(xClusterConfig)) {
-        xClusterConfig.setStatus(XClusterConfig.XClusterConfigStatusType.Deleted);
-      }
+      // Update DB to reflect this change.
+      xClusterConfig.setReplicationSetupDone(
+          xClusterConfig.getTables(), false /* replicationSetupDone */);
 
       if (HighAvailabilityConfig.get().isPresent()) {
         getUniverse(true).incrementVersion();
